+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - [**BC**] Change all DTO structures to PascalCase, and let the serialization dealt with a format
+- Add `StartProcessCommand` module
+- Add tests
 
 ## 1.1.0 - 2020-04-15
 - Make `CommonSerializer` public